ci: Linux ネイティブビルド (ncurses/CLI) を CI に載せる (#16 Phase 4)
src/frontend/ncurses と src/frontend/cli の 13,000 行は、MSVC のジョブでも
llvm-mingw のジョブでもコンパイルされていなかった。どちらも Windows 向けの
クロスビルドで、ncursesw が無い環境では xyzzy-ncurses ターゲットがそもそも
生成されないためで、結果として実際に腐っていた。ジョブを足すにあたって
その 2 件を直している。

- stdafx.h の min/max をマクロから using std::min/max へ。#33 で独自の
  min/max を std のものに統一した際にマクロだけが残っており、cdecl.h が
  <algorithm> を include する時点でマクロが生きている経路 (stdafx.h を先に
  読む .cc) で <algorithm> 自身がコンパイルエラーになる。llvm-mingw の
  libc++ は <list> の中で <algorithm> を読み切るので表面化せず、libstdc++
  でだけ壊れていた。消すのではなく using を置いたのは、privctrl 以下が
  cdecl.h を通らず非修飾の min/max をこのマクロだけに頼っていたため。型が
  揃っている必要が出るので、int と LONG が混ざる 3 箇所にキャストを入れた。

- ncurses-main.cc で user-config-path を初期化する。設定していたのは Win32
  の init.cc だけで、POSIX では #:unbound のまま。起動時に lisp/backup.l の
  (concat (user-config-path) ".xyzzy.d/backup/") がそれを掴んで startup.l の
  ロードごと落ちていた。つまり「リンクは通るが起動しない」状態だった。

ジョブはビルドに加えて tools/linux-smoke.sh を実行する。ライブラリを最後まで
読み込んで式を評価できるところまで見ないと、2 件目のような故障が素通りする。
Lisp テストスイートは Linux ビルドでは出力ゼロのままメモリを食い潰し続ける
(25 分で 12GB) ため対象外で、既知の課題として残す。

ローカルでの再現用に tools/x へ linux ターゲットと smoke を足し、devenv
イメージに libncurses-dev / zlib1g-dev を入れた。

+++ src/frontend/win32/privctrl/url.cc
@@ -55,7 +55,7 @@ dopaint (HWND hwnd, HDC hdc)
       l = wcslen (s[i]);
       SetTextColor (hdc, i == 1 ? hl : bg);
       GetTextExtentPoint32W (hdc, s[i], l, &sz);
-      r.right = min (r.left + sz.cx, xmax);
+      r.right = min (r.left + sz.cx, (LONG)xmax);
       ExtTextOutW (hdc, r.left, 0, ETO_CLIPPED, &r, s[i], l, 0);
       if (i == 1)
         {
